requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai>=1.107.2
pydantic==2.5.0
aiohttp>=3.10.0
pydantic-settings==2.1.0
python-socketio==5.10.0
python-engineio==4.8.0
requests>=2.31.0

# LiveKit Core
livekit==1.1.3
livekit-api==1.1.0
livekit-agents==1.5.1
livekit-plugins-openai==1.5.1
livekit-plugins-silero==1.5.1
livekit-plugins-hedra==0.1.0

# LiveKit Agent Dependencies
aiofiles>=24
av>=14.0.0
certifi>=2024.7.4
click~=8.1
colorama>=0.4.6
docstring-parser>=0.16
eval-type-backport
livekit-blingfire<2,~=1.1
livekit-protocol<2,>=1.1.3
nest-asyncio>=1.6.0
numpy>=1.26.0
onnxruntime>=1.18
prometheus-client>=0.22
psutil>=6.0.0
sounddevice>=0.5
typer>=0.15.1
watchfiles>=1.0
pillow>=10.3.0

# OpenTelemetry
opentelemetry-api~=1.39.0
opentelemetry-exporter-otlp~=1.39.0
opentelemetry-sdk~=1.39.0

# Additional Dependencies
pyjwt>=2.0.0
protobuf>=4
types-protobuf>=4
